avito_selenium_parser/requirements.txt
selenium
Pillow
pytesseract
bs4